refactor(Movie): ScreeningTimes에 있던 시간대 가져오는 로직을 Moive의 정적 함수로 이동
